Don't retry aborting multipart upload
Even though the AWS S3 documentation states that we should do it,
neither Uppy Companion nor tusd are doing this and nobody complained, so
apparently there is no need. If it was really that important, I would
expect aws-sdk-s3 to have a method for "safe aborting".

Since it's good to take every opportunity to make code simpler, we
remove the code that retries aborting the multipart upload.